Importer Texture split blender into own file
