+++ addons/io_scene_gltf2/blender/imp/material/__init__.py
@@ -21,5 +21,4 @@
  * This development is done in strong collaboration with Airbus Defence & Space
  """
  
-from .texture import *
 from .image import *
